[GH-31] updated battle menu to use animation speed from options
* refactored logic tie to setting animation speed from data manager to
no longer live in the dialog ui class, and to live in the data manager
class
* updated the battle menu class to use the new method in the data
manager class to set the delay speed for the text animation

GitHub Issue: https://github.com/devshareacademy/monster-tamer/issues/31
